move reaction menus code to separate trimatix package. Fix args being split on command call.
