--- requirements.txt
@@ -1,2 +1,0 @@
-discord.py
-emoji
