+++ src/requirements.txt
@@ -1,4 +1,5 @@
 psycopg2-binary>=2.8
 discord.py
 python-dotenv
+emoji
 git+git://github.com/JustAnotherArchivist/snscrape.git#egg=snscrape
