+++ src/trimatix/botState.py
@@ -1,0 +1,1 @@
+client = None
