+++ src/trimatix/lib/emotes.py
