+++ src/trimatix/lib/exceptions.py
